+++ apps/osrm-requester/src/coordinate.rs
@@ -1,16 +1,5 @@
 use serde::{Serialize, Deserialize};
-use std::fs;
 use geo_types::LineString;
-
-/// Конфиг
-#[derive(Serialize, Deserialize, Debug)]
-#[serde(rename_all = "kebab-case")]
-pub struct Config {
-    /// Очередь для запроса
-    pub request_queue: String,
-    /// Очередь для ответа
-    pub response_queue: String,
-}
 
 /// Координата
 #[derive(Serialize, Deserialize, Debug)]
@@ -20,14 +9,6 @@ pub struct Coordinate {
     latitude: f32,
     /// Долгота
     longitude: f32,
-}
-
-/// Получение конфига
-pub fn get_config() -> anyhow::Result<Config> {
-  let config_file_str: String = fs::read_to_string("config/config.yaml")?;
-  let config: Config = serde_yaml::from_str(&config_file_str)?;
-
-  Ok(config)
 }
 
 /// Получение строки координат (склеивание координат через запятую)
